% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fourier_DE.R
\name{fourier_est}
\alias{fourier_est}
\title{Fourier transformation function}
\usage{
fourier_est(y, t, p, h = c("1", "2", "3"))
}
\arguments{
\item{y}{Observed endogenous concentration}

\item{t}{Time array}

\item{p}{initial estimate of Fourier coefficient}

\item{h}{Number of harmonics. 3 is allowed}
}
\value{
this function provides estimation of fourier coefficients
}
\description{
Fourier series analysis will be done to fit the endogenous production mimicing the circadian rhythm
}
