% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_pred.R
\name{ft_pred}
\alias{ft_pred}
\title{Concentration prediction Fourier function}
\usage{
ft_pred(fourier_est, t, h)
}
\arguments{
\item{fourier_est}{function to estimate Fourier coefficients}

\item{t}{time data}
}
\description{
This function will allow to predict the concentration after the estimation from \verb{r fourier_est}
}
\author{
Mutaz J
}
