% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cosine.R
\name{single_cos}
\alias{single_cos}
\title{Single cosine model}
\usage{
single_cos(y, t, p)
}
\arguments{
\item{y}{Observed concentration}

\item{t}{time}

\item{p}{Initial values. p1: Amplitude p2: mean endogenous rate p3: acrophase time}
}
\description{
Estimation of Amp, mean concentration
}
