% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singe_cosine_pred.R
\name{single_cos_pred}
\alias{single_cos_pred}
\title{Single cosine function prediction}
\usage{
single_cos_pred(single_cos, t)
}
\arguments{
\item{single_cos}{function to predict single cosine coefficients}

\item{t}{time vector}
}
\description{
This function will pass \code{single_cos} arguments to predict the concentration
}
\author{
Mutaz J.
}
